test(github): complete GraphQL qualification integration
