+++ packages/@emulators/github/README.md
@@ -19,7 +19,7 @@ npm install @emulators/github
 - `Issue.comments` supports Relay-style `first`, `after`, `last`, and `before` arguments with opaque cursors.
 - Mutations support `createIssue`, `closeIssue`, `reopenIssue`, `addComment`, `createLabel`, `deleteLabel`, and exact `deleteIssue`; each accepts a typed input and echoes `clientMutationId` exactly. `deleteIssue` returns the owning repository, matching [GitHub's documented payload](https://docs.github.com/en/graphql/reference/issues#deleteissue), and cascades issue-owned comments, timeline records, labels, and relationship edges while preserving repository labels and unrelated records. No deletion webhook or issue event is emitted because this emulator has no evidence-backed provider behavior for one.
 - Requests require authentication and use the GraphQL rate-limit bucket exposed by `rateLimit` and `/rate_limit`.
-- The qualified subset is the three queries `repository`, `node`, and `rateLimit`, plus the nine issue-graph mutations `addSubIssue`, `addBlockedBy`, `createIssue`, `closeIssue`, `reopenIssue`, `addComment`, `createLabel`, `deleteLabel`, and `deleteIssue`. Connections use opaque cursors and 100-item pages. Seeded issue graphs and reset-specific guarantees remain pending issue #11 integration; no live GitHub calls are made.
+- The qualified subset is the three queries `repository`, `node`, and `rateLimit`, plus the nine issue-graph mutations `addSubIssue`, `addBlockedBy`, `createIssue`, `closeIssue`, `reopenIssue`, `addComment`, `createLabel`, `deleteLabel`, and `deleteIssue`. Connections use opaque cursors and 100-item pages. Seeded issue graphs and reset-specific guarantees are covered by qualification tests; no live GitHub calls are made.
 - Rate limits use one internally consistent local bucket shared by GraphQL and REST; GitHub operation-specific cost fidelity and remote quota enforcement are not emulated.
 - Issue relationships include `parent`, cursor-paginated `subIssues` and `blockedBy`, plus `addSubIssue` and `addBlockedBy` mutations with exact `clientMutationId` echoes.
